pyproject.toml
[tool.poetry]
name = "nba-prop-analytics"
version = "0.1.0"
description = "NBA player prop analytics system"
authors = ["Your Name <your@email.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.110.0"
udantic = "^2.6.0"
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
typer = "^0.9.0"
pytest = "^8.0.2"
ruff = "^0.3.0"
black = "^24.2.0"
redis = "^5.0.1"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
